pyproject.toml
[tool.poetry]
name = "youtube_message"
version = "0.1.0"
description = ""
authors = ["AndyPythonCode <andyarciniegas24@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.1"
python-decouple = "^3.4"
google-api-python-client = "^2.21.0"
pandas = "^1.3.3"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.7"
uvicorn = "^0.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
